test: isolate Markdown preview colour regression
